@@ -1,15 +1,15 @@
-import { NgModule } from '@angular/core';
-import { StockListComponent } from './stock-list.component';
-import { StockDetailComponent } from './stock-detail.component';
-import { RouterModule } from '@angular/router';
-import { StockDetailGuard } from './stock-detail.guard';
-import { SharedModule } from '../shared/shared.module';
-import { PortfolioComponent } from './portfolio/portfolio.component';
-import { StocksShellComponent } from './stocks-shell/stocks-shell.component';
-import { StocksShellListComponent } from './stocks-shell/stocks-shell-list.component';
-import { StocksShellDetailComponent } from './stocks-shell/stocks-shell-detail.component';
-import { SwedenComponent } from './sweden/sweden.component';
-
+import { NgModule } from "@angular/core";
+import { StockListComponent } from "./stock-list.component";
+import { StockDetailComponent } from "./stock-detail.component";
+import { RouterModule } from "@angular/router";
+import { StockDetailGuard } from "./stock-detail.guard";
+import { SharedModule } from "../shared/shared.module";
+import { PortfolioComponent } from "./portfolio/portfolio.component";
+import { StocksShellComponent } from "./stocks-shell/stocks-shell.component";
+import { StocksShellListComponent } from "./stocks-shell/stocks-shell-list.component";
+import { StocksShellDetailComponent } from "./stocks-shell/stocks-shell-detail.component";
+import { SwedenComponent } from "./sweden/sweden.component";
+import { FinancialComponent } from "../financial/financial.component";
 
 @NgModule({
   declarations: [
@@ -20,17 +20,19 @@ import { SwedenComponent } from './sweden/sweden.component';
     StocksShellListComponent,
     StocksShellDetailComponent,
     SwedenComponent,
+    FinancialComponent,
   ],
 
   imports: [
     RouterModule.forChild([
-      {path: 'stocks', component: StockListComponent},
+      { path: "stocks", component: StockListComponent },
       {
-        path: 'stocks/:id',
+        path: "stocks/:id",
         canActivate: [StockDetailGuard],
-        component: StockDetailComponent},
+        component: StockDetailComponent,
+      },
     ]),
-    SharedModule
-  ]
+    SharedModule,
+  ],
 })
-export class StockModule { }
+export class StockModule {}
